+++ src/Akeneo/Tool/Bundle/MeasureBundle/tests/EndToEnd/InternalApi/CreateMeasurementFamilyActionEndToEnd.php
@@ -12,7 +12,7 @@ use Akeneo\Tool\Bundle\MeasureBundle\Persistence\MeasurementFamilyRepositoryInte
 use Akeneo\Tool\Bundle\MeasureBundle\tests\EndToEnd\WebTestCase;
 use Symfony\Component\HttpFoundation\Response;
 
-class CreateMeasurementFamilyActionTest extends WebTestCase
+class CreateMeasurementFamilyActionEndToEnd extends WebTestCase
 {
     /** @var MeasurementFamilyRepositoryInterface */
     private $measurementFamilyRepository;
@@ -79,12 +79,11 @@ class CreateMeasurementFamilyActionTest extends WebTestCase
         );
 
         $response = $this->client->getResponse();
-        $this->assertSame(Response::HTTP_UNPROCESSABLE_ENTITY, $response->getStatusCode());
+        $this->assertSame(Response::HTTP_BAD_REQUEST, $response->getStatusCode());
         $responseBody = json_decode($response->getContent(), true);
-        $this->assertEquals(Response::HTTP_UNPROCESSABLE_ENTITY, $responseBody['code']);
         $this->assertEquals(
-            'The measurement family has data that does not comply with the business rules.',
-            $responseBody['message']
+            'This field can only contain letters, numbers, and underscores.',
+            $responseBody[0]['message']
         );
     }
 
@@ -123,7 +122,7 @@ class CreateMeasurementFamilyActionTest extends WebTestCase
         );
 
         $response = $this->client->getResponse();
-        $this->assertSame(Response::HTTP_UNPROCESSABLE_ENTITY, $response->getStatusCode());
+        $this->assertSame(Response::HTTP_BAD_REQUEST, $response->getStatusCode());
     }
 
     private function measurementFamilyWithCode(string $code): MeasurementFamily