+++ CHANGELOG.md
@@ -59,6 +59,7 @@
 - PIM-9829: Fix product grid crash when using a family filter on a deleted family
 - PIM-9820: Fix the Error 500 on the product grid with the date filter
 - PIM-9833: Fix null pointer exception on Product::getVariationLevel (CE contribution)
+- PIM-9826: Display the system attribute filters with the UI locale on the user account settings
 
 ## New features